move model capabilities into providers
